crawler/requirements.txt
# HTTP client
httpx>=0.27.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Structured logging
structlog>=24.0.0

# Retry logic
tenacity>=8.2.0

# Progress display
tqdm>=4.66.0

# Data validation
pydantic>=2.5.0

# Async file I/O
aiofiles>=23.2.0

# Environment variables
python-dotenv>=1.0.0

# Neon / Postgres client (replaces supabase SDK)
# psycopg2-binary includes the C extension — no libpq build required
psycopg2-binary>=2.9.9

# Primary embeddings — local, no API key needed
# Model: BAAI/bge-small-en-v1.5, 384 dimensions
fastembed>=0.3.0

# Groq — primary LLM for thread summarization (free tier, no cost)
groq>=0.9.0

# OpenAI client — only needed when EMBEDDING_PROVIDER=openai
openai>=1.10.0

# Token counting for OpenAI embedding truncation
tiktoken>=0.6.0
